Introduces argument to control redirect follow count
This commit introduces an argument that let you limit the amount of
redirect follows the client will do, before giving up and returning the
response.

Right now, the default is set to 5.

This changed required changes into the `hyper::client` module. It has
been proposed on a PR[1], but yet to be merged. There are some plans to
move into async strategies, which will require a rewrite of this
anyways.

Meanwhile, this crate will depend on the fork located on
[bltavares/hyper](https://github.com/bltavares/hyper), which contains
the logic to do redirects using a provided policy.

[1] https://github.com/hyperium/hyper/pull/759